feature: Move common components outside, Add chatbubble component
