+++ src/components/ChatBubble/index.tsx
@@ -1,0 +1,7 @@
+import React, { PropsWithChildren } from 'react';
+
+const ChatBubble = (): JSX.Element => {
+  return <div></div>;
+};
+
+export default ChatBubble;
